% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_radius_mcmc.R
\name{adaptive_radius_MCMC}
\alias{adaptive_radius_MCMC}
\title{Local DTA MCMC (Adaptive Radius)}
\usage{
adaptive_radius_MCMC(
  N,
  ED,
  coal_rate,
  bit_mig_mat,
  st_radius,
  adaptive = TRUE,
  cr_mode,
  cr_var,
  mm_mode,
  mm_var,
  thin = max(N/5000, 1),
  save_trees = TRUE,
  tree_thin = thin,
  proposal_rates = c(1000, 1, 1),
  adaptation_rate = 0.6,
  target_accept_rate = 0.234,
  output_dir,
  run_name = "Local_DTA"
)
}
\arguments{
\item{N}{Total number of MCMC iterations to complete}

\item{ED}{Migration history to begin at in ED format (9 columns)}

\item{coal_rate}{Initial estimate of coalescent rates}

\item{bit_mig_mat}{Initial estimate of backward-in-time migration rates matrix}

\item{st_radius}{Initial radius for subtrees}

\item{adaptive}{Logical value indicating whether or not to adapt the radius within the run (default TRUE)}

\item{cr_mode}{Coalescent rates prior mode}

\item{cr_var}{Coalescent rates prior variance}

\item{mm_mode}{Backward-in-time migration rates prior mode}

\item{mm_var}{Backward-in-time migration rates prior variance}

\item{thin}{Thinning rate for continuous parameter posterior samples}

\item{save_trees}{Logical value indicating whether tree samples are saved (i.e. whether .trees file is created)}

\item{tree_thin}{Thinning rate for tree samples to be saved}

\item{proposal_rates}{Relative rates of migration history and continuous parameter updates (migration history : coalescent rates : migration rates)}

\item{adaptation_rate}{Rate at which adaptive MCMC varies subtree radius (unused if adaptive = FALSE)}

\item{target_accept_rate}{Target acceptance rate for migration history updates}

\item{output_dir}{Directory to output log files to}

\item{run_name}{Run name to save log files as}
}
\description{
Runs an MCMC chain using the local DTA method with radius-based subtrees (adaptively set radius)
}
