% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTree.R
\name{approximate_consensus_tree}
\alias{approximate_consensus_tree}
\title{Consensus tree}
\usage{
approximate_consensus_tree(
  ED_list,
  n_splits = 5,
  consensus_prob = 0.5,
  plot = TRUE
)
}
\arguments{
\item{n_splits}{Number of splits per branch used to discretise the tree}

\item{consensus_prob}{Consensus probability to be exceeded to allow colouring}

\item{plot}{Logical value whether to plot tree with superimposed pie charts}
}
\value{
Returns consensus tree in ED format
}
\description{
Generates an approximate consensus tree by discretising branches and colouring each section if the consensus probability is sufficiently large
}
