% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_conversion.R
\name{as.ED}
\alias{as.ED}
\alias{as.EED}
\alias{as.ED.default}
\alias{as.EED.default}
\alias{as.ED.phylo}
\alias{as.ED.EED}
\alias{as.EED.ED}
\alias{as.EED.phylo}
\alias{as.phylo.ED}
\alias{as.phylo.EED}
\title{Conversion Between Tree Objects}
\usage{
as.ED(x)

as.EED(x)

\method{as.ED}{default}(x, ...)

\method{as.EED}{default}(x, ...)

\method{as.ED}{phylo}(x, ...)

\method{as.ED}{EED}(x, ...)

\method{as.EED}{ED}(x, ...)

\method{as.EED}{phylo}(x, ...)

\method{as.phylo}{ED}(x, ...)

\method{as.phylo}{EED}(x, ...)
}
\arguments{
\item{x}{object to be converted}
}
\value{
An object of class "ED"
}
\description{
as.ED is a generic function converting a tree object into a tree of class "ED".

as.EED is a generic function converting a tree object into a tree of class "EED".

There are currently methods implemented to convert between objects of type "phylo", "ED", "EED"
}
