% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_conversion.R
\name{as.strphylo}
\alias{as.strphylo}
\alias{as.strphylo.default}
\alias{as.strphylo.phylo}
\title{Conversion to Structured Phylo}
\usage{
as.strphylo(x, ...)

\method{as.strphylo}{default}(x, ...)

\method{as.strphylo}{phylo}(x, node.deme, ...)
}
\arguments{
\item{x}{object to be converted}
}
\value{
An object of class "strphylo"
}
\description{
Conversion to structured phylo objects
}
