% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_sampling.R
\name{bit_global_DTA}
\alias{bit_global_DTA}
\title{DTA Sampling}
\usage{
bit_global_DTA(ED, bit_rates, root_deme = NA)
}
\arguments{
\item{ED}{Extended data representation of a phylogeny including migration history}

\item{bit_rates}{Rates matrix for the backwards-in-time migration process (diagonal elements = - sum of off-diagonal elements)}

\item{root_deme}{Specify the deme to be selected at the root of the tree}
}
\value{
List of proposed DTA migration histories with the same leaves as input ED
}
\description{
Samples N migration histories under the DTA model deme fixed at all leaves using belief propagation
}
