% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BitMigMatC}
\alias{BitMigMatC}
\title{Backward-in-time migration matrix conversion}
\usage{
BitMigMatC(fit_mm, coal_rate)
}
\arguments{
\item{coal_rate}{NumericVector Coalescent rates}

\item{bit_mm}{NumericMatrix Forward-in-time migration matrix}
}
\value{
NumericMatrix Backward-in-time migration matrix
}
\description{
Computes the backward-in-time migration matrix associated with a given forward-in-time migration matrix, coalescent rate and time scale
}
