% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{coal.split}
\alias{coal.split}
\title{Coalescent Node Split}
\usage{
coal.split(edge, edge.deme, node.ages, n.deme)
}
\arguments{
\item{edge}{Edge matrix for a structured coalescent tree}

\item{edge.deme}{Label for the deme in which each edge lies}

\item{node.ages}{Time of each node in the coalescent tree}

\item{n.deme}{Number of possible demes in the process}
}
\value{
List containing the updated edge matrix,
node ages and edge demes
}
\description{
Performs a coalescent node split (Ewing et al. 2004). Selects a coalescent
node uniformly at random and if the parent of the coalescent node is a
migration node, the migration node is pulled through the coalescent node onto
the child edges.
}
