% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_DTA.R
\name{coal_rate_gibbs_update}
\alias{coal_rate_gibbs_update}
\alias{mig_mat_gibbs_update}
\title{Gibbs Updates}
\usage{
coal_rate_gibbs_update(ED, ED_NI, n_deme, prior_shape = 1, prior_rate = 1)

mig_mat_gibbs_update(ED, n_deme, ED_NI, prior_shape = 1, prior_rate = 1)
}
\arguments{
\item{ED}{Extended data representation of a phylogeny including migration history}

\item{ED_NI}{Vector identifying row of each node label}

\item{n_deme}{Number of demes used in structured coalescent model}

\item{prior_shape}{Gamma shape parameter for prior}

\item{prior_rate}{Gamma rate parameter for prior}
}
\value{
Updated parameters
}
\description{
Performs Gibbs updates for structured coalescent parameters using conjugate gamma prior distributions
}
