% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coal_node_pie_charts.R
\name{coalescent_node_pie_charts}
\alias{coalescent_node_pie_charts}
\title{Coalescent Node Pie Charts}
\usage{
coalescent_node_pie_charts(ED_list, plot = TRUE, plot_ED = matrix(NA, 0, 9))
}
\arguments{
\item{plot}{Logical value whether to plot tree with superimposed pie charts}

\item{plot_ED}{(optional) ED to superimpose with pie charts (clear migration history if left blank)}

\item{ED_sample}{Sample of migration histories on same topology (ED with 9 columns)}
}
\value{
Returns deme frequencies at each location
}
\description{
Tabulates the frequency of each deme at each leaf and coalescent node of a phylogenetic tree
}
