% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTree.R
\name{consensus_tree}
\alias{consensus_tree}
\title{Consensus tree}
\usage{
consensus_tree(
  ED_sample,
  n_splits = 5,
  consensus_prob = 0.8,
  n_deme = NA,
  plot = TRUE
)
}
\arguments{
\item{ED_sample}{Sample of migration histories on same topology}

\item{n_splits}{Number of splits per branch used to discretise the tree}

\item{consensus_prob}{Consensus probability to be exceeded to allow colouring}

\item{n_deme}{(optional) Number of demes}
}
\value{
Returns deme frequencies at each location
}
\description{
Generates a consensus tree by discretising branches and colouring each section if the consensus probability is sufficiently large
}
