% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DemeDecompC}
\alias{DemeDecompC}
\title{Deme Decomposition}
\usage{
DemeDecompC(ED, n_deme, node_indices)
}
\arguments{
\item{ED}{Extended data representation of a migration history}

\item{n_deme}{Number of demes}

\item{node_indices}{Vector with element i giving row index of label i (NA if not present)}
}
\value{
m Matrix giving number of each type of migration observed

c Vector giving number of coalescences per deme
}
\description{
Counts the number of each type of migration event and deme of each coalescence event
}
