% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TruncatedExp.R
\name{dexp.trunc}
\alias{dexp.trunc}
\alias{pexp.trunc}
\alias{rexp.trunc}
\title{The Truncated Exponential Distribution}
\usage{
dexp.trunc(x, c = Inf, rate = 1)

pexp.trunc(q, c = Inf, rate = 1)

rexp.trunc(n, c = Inf, rate = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{c}{truncation cutoff}

\item{rate}{vector of rates}

\item{n}{number of observations}
}
\description{
Density, distribution function and random generation for the exponential
distribution with rate \code{rate} truncated above by \code{c}
}
