% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihoods.R
\name{dta.likelihood}
\alias{dta.likelihood}
\title{Calculates the DTA likelihood for a structured coalescent tree}
\usage{
dta.likelihood(
  ED,
  effective_population,
  gen_length,
  migration_matrix,
  node_indices
)
}
\arguments{
\item{ED}{Extended data representation of a structured coalescent process
@param effective_population effective population sizes from which the samples are taken}

\item{gen_length}{mean generation length of the sampled individuals}

\item{migration_matrix}{matrix of migration rates between demes}

\item{node_indices}{Vector giving row indices for each node label}
}
\value{
A vector (\code{log-likelihood}, \code{likelihood}) giving the log-likelihood and likelihood of \code{phylo}
}
\description{
Returns the DTA likelihood and log-DTA-likelihood of a structured coalescent tree
}
