% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihoods.R, R/Local_likelihoods.R
\name{DTA_likelihood}
\alias{DTA_likelihood}
\alias{DTA_local_likelihood}
\title{Discrete Trait Analysis likelihood}
\usage{
DTA_likelihood(ED, fit_mig_mat, ED_NI)

DTA_local_likelihood(st_labels, coal_rate, bit_mig_mat)
}
\arguments{
\item{ED}{Extended data representation of a structured genealogy}

\item{fit_mig_mat}{Matrix of forwards-in-time migration rates}

\item{ED_NI}{Node indices obtained from NodeIndices()}

\item{st_labels}{Reduced extended data structure consisting of the rows of ED corresponding to nodes in the selected subtree only}

\item{coal_rate}{Vector of coalescent rates}

\item{bit_mig_mat}{Matrix of backwards-in-time migration rates}
}
\value{
A vector (\code{log-likelihood}, \code{likelihood}) giving the log-likelihood and likelihood of \code{phylo}
}
\description{
Returns the DTA likelihood and log-DTA-likelihood of a structured coalescent tree
}
