% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_sampling.R
\name{DTA_sampling}
\alias{DTA_sampling}
\title{DTA Sampling}
\usage{
DTA_sampling(
  ED,
  mig_mat,
  time_scale = 1,
  N = 1,
  parallel = FALSE,
  mc.cores = NA
)
}
\arguments{
\item{ED}{Extended data representation of a phylogeny including migration history}

\item{time_scale}{Time scale for the migration matrix}

\item{N}{Number of migration histories to sample}

\item{parallel}{Logical, whether or not to generate proposals in parallel}

\item{mc.cores}{(optional) If parallel is TRUE, number of cores to parallelise over}

\item{fit_mig_mat}{Forward-in-time migration matrix for the phylogeny}
}
\value{
List of proposed DTA migration histories with the same leaves as input ED
}
\description{
Samples N migration histories under the DTA model deme fixed at all leaves using belief propagation
}
