% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDBlockRecolour.R
\name{ed.block.recolour}
\alias{ed.block.recolour}
\title{Block Recolouring MCMC Proposal}
\usage{
ed.block.recolour(ED, n.deme, fix.leaf.deme = TRUE, node.indices)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n.deme}{Number of distinct demes in the population}

\item{fix.leaf.deme}{Logical; if TRUE, the deme of each leaf in the tree is fixed. Any proposal which attempts to change a leaf deme is automatically rejected}
}
\value{
Updated extended data object with the proposal from the block recolouring proposal
}
\description{
Performs a block recolouring MCMC proposal on a structured coalescent
migration history. A block of the migration history currently in a single
deme is migrated into a new deme selected uniformly from all other demes. If
an inconsistent migration history is obtained, the move will be rejected.
}
