% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ewing_ED_proposals.R
\name{ed.coal.split}
\alias{ed.coal.split}
\title{Coalescent Node Split Proposal}
\usage{
ed.coal.split(ED, n.deme, node.indices)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n.deme}{Number of distinct demes in the population}

\item{node.indices}{Vector giving row indices for node labels}
}
\value{
Updated extended data object with the proposal from the migration pair birth move
}
\description{
Performs a coalescent node split move (Ewing et al. 2004). Splits a migration
node immediately above a coalescent node into two migration nodes placed immediately
below the coalescent node
}
