% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDDistance.R
\name{ED.dist.coal.node}
\alias{ED.dist.coal.node}
\title{ED Distance (coalescent node)}
\usage{
ED.dist.coal.node(
  ED1,
  ED2,
  mm_1,
  mm_2,
  node_indices_1 = NA,
  node_indices_2 = NA
)

ED.dist.coal.node(
  ED1,
  ED2,
  mm_1,
  mm_2,
  node_indices_1 = NA,
  node_indices_2 = NA
)
}
\arguments{
\item{ED1}{Extended data representation of first migration history}

\item{ED2}{Extended data representation of second migration history}

\item{mm_1}{Migration matrix for ED1}

\item{mm_2}{Migration matrix for ED2}
}
\value{
Computed distance

Computed distance
}
\description{
Computes the dissimilarity between a pair of migration histories on the same
tree via number of coalescent nodes in different demes between the two histories

Computes the dissimilarity between a pair of migration histories on the same
tree via number of coalescent nodes in different demes between the two histories
}
