% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDDistance.R
\name{ED.dist}
\alias{ED.dist}
\title{ED Distance}
\usage{
ED.dist(ED1, ED2, node_indices_1 = NA, node_indices_2 = NA)
}
\arguments{
\item{ED1}{Extended data representation of first migration history}

\item{ED2}{Extended data representation of second migration history}
}
\value{
Computed distance
}
\description{
Computes the dissimilarity between a pair of migration histories on the same
tree via total branch length in different demes
}
