% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloLikelihood.R
\name{ed.likelihood}
\alias{ed.likelihood}
\title{Calculates the likelihood of a structured coalescent tree}
\usage{
ed.likelihood(ED, effective.pop, gen.length, migration.matrix, node.indices)
}
\arguments{
\item{ED}{Extended data representation of a structured coalescent process
@param effective.pop effective population sizes from which the samples are taken}

\item{gen.length}{mean generation length of the sampled individuals}

\item{migration.matrix}{matrix of migration rates between demes}
}
\value{
A vector (\code{log-likelihood}, \code{likelihood}) giving the log-likelihood and likelihood of \code{phylo}
}
\description{
Returns the likelihood and log-likelihood of a tree generated under the
structured coalescent process
}
