% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMigBirthDeathNEW.R
\name{ed.mig.birth.new}
\alias{ed.mig.birth.new}
\title{Migration Birth MCMC Move}
\usage{
ed.mig.birth.new(ED, n.deme)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n.deme}{Number of distinct demes in the population}
}
\value{
Updated extended data object with the proposal from the migration birth move
}
\description{
Selections a location uniformly over a structured coalescent genealogy and
attempts to add a migration node, reallocating demes to obtain a consistent
migration history
}
