% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ewing_ED_proposals.R
\name{ed.mig.birth}
\alias{ed.mig.birth}
\title{Migration Birth MCMC Move}
\usage{
ed.mig.birth(ED, n.deme, fix.leaf.deme = TRUE, node.indices)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n.deme}{Number of distinct demes in the population}
}
\value{
Updated extended data object with the proposal from the migration birth move
}
\description{
Performs a Migration birth move (Ewing et al. 2004). Adds a migration
node between a randomly selected ancestral node and its parent, allocating a
deme for the new edge consistent with the surrounding edges.
}
