% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMigBirthDeathNEW.R
\name{ed.mig.death.new}
\alias{ed.mig.death.new}
\title{Migration Death MCMC Move}
\usage{
ed.mig.death.new(ED, n.deme)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n.deme}{Number of distinct demes in the population}
}
\value{
Updated extended data object with the proposal from the migration death move
}
\description{
Performs a Migration death move. Selects a migration event
uniformly at random and removes the parent of the selected node, pulling the
deme from the parent edge downwards if a consistent deme labelling can be
obtained.
}
