% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ewing_ED_proposals.R
\name{ed.mig.death}
\alias{ed.mig.death}
\title{Migration Death MCMC Move}
\usage{
ed.mig.death(ED, n.deme, fix.leaf.deme = TRUE, node.indices)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n.deme}{Number of distinct demes in the population}
}
\value{
Updated extended data object with the proposal from the migration death move
}
\description{
Performs a Migration death move (Ewing et al. 2004). Selects an ancestral node
uniformly at random and removes the parent of the selected node, allocating
demes as necessary for modified edges.
}
