% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMigPairBirthDeath.R
\name{ed.mig.pair.birth}
\alias{ed.mig.pair.birth}
\title{Migration Pair Birth MCMC Move}
\usage{
ed.mig.pair.birth(ED, n.deme)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n.deme}{Number of distinct demes in the population}
}
\value{
Updated extended data object with the proposal from the migration pair birth move
}
\description{
Performs a Migration pair birth move (Ewing et al. 2004). Adds two migration
nodes on an edge selected uniformly at random from a structured coalescent
process, allocating a deme for the added edge such that a migration event
does not target its origin deme.
}
