% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMigPairBirthDeath.R
\name{ed.mig.pair.death}
\alias{ed.mig.pair.death}
\title{Migration Pair Death MCMC Move}
\usage{
ed.mig.pair.death(ED, n.deme)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n.deme}{Number of distinct demes in the population}
}
\value{
Updated extended data object with the proposal from the migration pair death move
}
\description{
Performs a Migration pair death move (Ewing et al. 2004). Deletes two
migration nodes if they lie between two edges in the same deme.
}
