% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED_migration_shift.R
\name{ed.mig.shift.1}
\alias{ed.mig.shift.1}
\title{Migration Shift MCMC Proposal}
\usage{
ed.mig.shift.1(ED, n.deme)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n.deme}{Number of distinct demes in the population}
}
\value{
Updated extended data object with the proposal from the migration shift proposal
}
\description{
Performs a migration shift MCMC proposal on a structured coalescent
migration history. A migration event is selected uniformly at random
and the time of the event is updated by resampling the event time between
the times of the preceding and following events.
}
