% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeCount.R
\name{ed.node.count}
\alias{ed.node.count}
\title{Counting of Migration and Coalescent Events}
\usage{
ed.node.count(ED, n.deme, node.indices)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n.deme}{Number of demes in the structured coalescent process (optional argument)}
}
\value{
Returns a list containing a matrix M and a vector C. M has entries {m_{ij}} giving the number of migrations beginning in deme i and ending in deme j backwards in time. C has entries c_i giving the number of coalescence events occurring in deme i.
}
\description{
Counts the number of migration events between each pair of demes and the number
of coalescence events occurring in each deme
}
