% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED_conversion.R
\name{ed.to.eed}
\alias{ed.to.eed}
\title{ED to EED Conversion}
\usage{
ed.to.eed(ED, node_indices = NA)
}
\arguments{
\item{ED}{phylo object augmented with the deme of each node}

\item{node_indices}{(optional) vector relating row labels with row number}
}
\value{
A matrix with each row containing the parent node, child nodes, deme,
age, parent coalescent node and child coalescent nodes of
each node in the tree.
}
\description{
Converts an EED object for a structured coalescent tree into an EED, adding
the parent and children coalescent nodes for each node
}
