% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedDataConversion.R
\name{ed.to.phylo}
\alias{ed.to.phylo}
\title{Extended Data to Phylo Object}
\usage{
ed.to.phylo(ED)
}
\arguments{
\item{ED}{Extended data object to convert; matrix with columns Node ID, parent, child 1, child 2, deme, node age}
}
\value{
A phylo object augmented with node demes.
}
\description{
Converts an extended data object into a phylo object augmented with the deme
of each node
}
