% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ED_dist_C}
\alias{ED_dist_C}
\title{ED_Dist}
\usage{
ED_dist_C(ED1, ED2, n_deme, node_indices_1 = NULL, node_indices_2 = NULL)
}
\arguments{
\item{ED1}{NumericMatrix Extended data object representing first migration history}

\item{ED2}{NumericMatrix Extended data object representing second migration history}

\item{node_indices_1}{NumericVector Vector giving row numbers in ED1 for node labels}

\item{node_indices_2}{NumericVector Vector giving row numbers in ED1 for node labels}
}
\value{
double Proportion of tree in different demes between ED1 and ED2
}
\description{
Computes the distance between pairs of migration histories on a common topology via proportion of branch lengths different between the two
}
