% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_sampling.R
\name{EED_local_DTA}
\alias{EED_local_DTA}
\title{Local DTA Update for EED data structure}
\usage{
EED_local_DTA(EED, fit_mig_mat, time_scale, selected_node = NA)
}
\arguments{
\item{EED}{EED representation of a phylogeny including migration history}

\item{time_scale}{Time scale for the migration matrix}

\item{selected_node}{Label for node proposal will be centered on (optional)}

\item{mig_mat}{Forward-in-time (relative) migration matrix for the phylogeny}
}
\value{
Proposal ED
}
\description{
Updates a migration history using the DTA model for adjacent branches to a coalescent node
}
