% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffPopUpdate.R
\name{eff.pop.update}
\alias{eff.pop.update}
\title{Effective Population Size Update}
\usage{
eff.pop.update(ED, effective.population, n.deme, alpha = 1, beta = 1)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{effective.population}{Current migration rates matrix for the structured coalescent process}

\item{n.deme}{Number of demes in the structured coalescent process (optional argument)}

\item{alpha}{Inverse-gamma shape parameter for the prior on effective population sizes}

\item{beta}{Inverse-gamma rate parameter for the prior on effective population sizes}
}
\value{
Updated vector of effective population sizes conditioned on the current migration history
}
\description{
Performs an update of the effective population sizes via a Gibbs move with Inverse-gamma prior and Inverse-gamma proposal distribution
}
