% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactConsensus.R
\name{exact_consensus}
\alias{exact_consensus}
\title{Majority Consensus Migration History}
\usage{
exact_consensus(ED_list, consensus_prob = 0.5, plot = TRUE, ...)
}
\arguments{
\item{ED_list}{Sample of migration histories on same topology}

\item{consensus_prob}{Consensus probability (must be greater than 0.5 to guarantee unique consensus history)}

\item{plot}{Logical value whether to output plot of consensus migration history}

\item{...}{Additional arguments to pass to plot}
}
\value{
Consensus migration history in ED format
}
\description{
Computes a majority consensus migration history for a sample of structured phylogenies on the same underlying phylogeny
}
