% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_scaled_MCMC.R
\name{fit_scaled_MCMC}
\alias{fit_scaled_MCMC}
\title{Runs an MCMC on Migration Histories}
\usage{
fit_scaled_MCMC(
  N0 = 1e+05,
  N = 1e+06,
  ED,
  coal_rate,
  time_scale,
  mig_mat,
  n_deme = NA,
  proposal.rates = c(rep(5, 4), 1, 1, 1),
  cr_prior_shape = 1,
  cr_prior_rate = 1,
  mm_prior_shape = 1,
  mm_prior_rate = 1,
  likelihood = "Structured",
  output.directory = "./MCMC_Results",
  create.new.directory = TRUE
)
}
\arguments{
\item{N0}{Number of iterations to use as a burn-in}

\item{leaf.deme}{vector of demes for each leaf in \code{phylo}}
}
\value{
An object of class \code{phylo} augmented with the deme of each node in the tree.
}
\description{
Runs an MCMC to infer migration histories under the structured coalescent model
}
