% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitch.R
\name{fitch}
\alias{fitch}
\title{Fitch Algorithm}
\usage{
fitch(ED, node_indices)
}
\arguments{
\item{ED}{Structured phylogeny giving structured tree with labelled tip demes}

\item{node_indices}{Output from NodeIndicesC()}
}
\description{
Computes the minimum number of migration events required to explain a structured phylogeny
}
