% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitMigMatC}
\alias{FitMigMatC}
\title{Forward-in-time migration matrix conversion}
\usage{
FitMigMatC(bit_mm, coal_rate)
}
\arguments{
\item{bit_mm}{NumericMatrix Backward-in-time migration matrix}

\item{coal_rate}{NumericVector Coalescent rates}
}
\value{
NumericMatrix Forward-in-time migration matrix
}
\description{
Computes the forward-in-time migration matrix associated with a given backwards-in-time migration matrix, coalescent rate and time scale
}
