% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_migration_matrix.R
\name{forward.migration.matrix}
\alias{forward.migration.matrix}
\title{Forward-in-time Migration Matrix}
\usage{
forward.migration.matrix(
  migration.matrix,
  effective.population = rep(1, dim(migration.matrix)[1])
)
}
\arguments{
\item{migration.matrix}{Backwards-in-time migration matrix}

\item{effective.population}{Vector of effective population sizes}
}
\value{
Forward-in-time migration matrix
}
\description{
Computes a forward-in-time migration matrix given a backwards-in-time migration matrix and effective population sizes
}
