% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get.edge.id}
\alias{get.edge.id}
\title{Get edge ID}
\usage{
get.edge.id(start, end, edge)
}
\arguments{
\item{start}{First endpoint of the edge}

\item{end}{Second endpoint of the edge}

\item{edge}{Edge matrix from a \code{phylo} object}
}
\value{
Edge ID for the edge connecting \code{start} and \code{end}
}
\description{
Returns edge ID in an edge matrix from a \code{phylo} object given two nodes
}
