% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeterochronousSim.R
\name{Heterochronous.Sim}
\alias{Heterochronous.Sim}
\title{Simulation of Heterochronous Coalescent Trees}
\usage{
Heterochronous.Sim(Data, effective.pop, gen.length, phylo.plot = FALSE)
}
\arguments{
\item{Data}{nx2 matrix with first column giving the tip labels and second column the time at which sample was taken}

\item{phylo.plot}{logical; if FALSE (default) plot is not produced}

\item{EffectivePop}{effective population size from which the sample is taken}

\item{GenLength}{generation length of the sampled individuals}
}
\value{
An object of class \code{phylo} (from package \code{ape}) augmented with the likelihood and log-likelihood of the simulated tree
}
\description{
Simulates a coalescent tree by estimating coalescence times backwards in time
for a heterochronous sample taken at different time.
}
