% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phylo_simulation.R
\name{Homochronous.Sim}
\alias{Homochronous.Sim}
\title{Simulation of Homochronous Coalescent Trees}
\usage{
Homochronous.Sim(Data, EffectivePop, GenLength)
}
\arguments{
\item{Data}{vector of labels to be assigned to the tips}

\item{EffectivePop}{effective population size from which the sample is taken}

\item{GenLength}{generation length of the sampled individuals}
}
\value{
An object of class \code{phylo} (from package \code{ape})
}
\description{
Simulates a coalescent tree by estimating coalescence times backwards in time
for a homochronous sample taken all at one time.
}
