% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initial_tree.R
\name{initial.ed}
\alias{initial.ed}
\title{Generates an initial tree for MCMC}
\usage{
initial.ed(ED, leaf.deme)
}
\arguments{
\item{ED}{object of class \code{ED} giving coalescent node topology}

\item{leaf.deme}{vector of demes for each leaf in \code{phylo}}
}
\value{
An object of class \code{phylo} augmented with the deme of each node in the tree.
}
\description{
Adds migration nodes to a structured coalescent topology to create a full
structured coalescent tree. Distribution of demes at each coalescent node
}
\details{
Generates an initial migration history for a structured coalescent tree given demes of every tip using ancestral character evolution (ACE)
}
