% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialTree.R
\name{initial.tree}
\alias{initial.tree}
\title{Generates an initial tree for MCMC}
\usage{
initial.tree(phylo, leaf.deme)
}
\arguments{
\item{phylo}{object of class \code{phylo} giving coalescent node topology}

\item{leaf.deme}{vector of demes for each leaf in \code{phylo}}
}
\value{
An object of class \code{phylo} augmented with the deme of each node in the tree.
}
\description{
Adds migration nodes to a structured coalescent topology to create a full
structured coalescent tree. Distribution of demes at each coalescent node
}
\details{
Generates an initial migration history for a structured coalescent tree given demes of every tip using ancestral character evolution (ACE)
}
