% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_DTA.R
\name{local_DTA}
\alias{local_DTA}
\title{Local DTA Sampling}
\usage{
local_DTA(
  ED,
  st_labels,
  fit_rates,
  ED_NI = NodeIndices(ED),
  eigen_decomp = eigen(fit_rates),
  inverse_vecs = solve(eigen_decomp$vectors)
)
}
\arguments{
\item{ED}{Extended data representation of phylogenetic tree and initial migration history}

\item{st_labels}{Reduced extended data structure consisting of the rows of ED corresponding to coalescent nodes in the selected subtree only}

\item{ED_NI}{Vector of row indices corresponding to which row of ED corresponds to each node label}

\item{bit_rates}{Transition matrix of a Markov process corresponding to a process with rates given by the backwards-in-time migration matrix}
}
\value{
List consisting of proposal, a structured phylogeny with updates made to the selected subtree, and prop_prob, the probability of the given update
}
\description{
Updates the migration history associated with a subtree under a conditional DTA model
}
