% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_local_DTA.R
\name{local_DTA_mcmc}
\alias{local_DTA_mcmc}
\title{DTA Sampling}
\usage{
local_DTA_mcmc(
  N = 1e+06,
  ED,
  coal_rate,
  bit_mig_mat,
  fit_mig_mat = FitMigMatC(bit_mig_mat, coal_rate),
  cr_mode = 1,
  cr_var = 1,
  mm_mode = 0.05,
  mm_var = 0.5,
  output_dir = "~",
  run_name = "Local_DTA",
  thin = 1000,
  prop_rates = c(400, 1, 1)
)
}
\arguments{
\item{N}{Number of MCMC iterations (including burn-in)}

\item{ED}{Extended data representation of a phylogeny including migration history}

\item{coal_rate}{Vector of coalescent rates in each deme}

\item{bit_mig_mat}{Matrix of backwards-in-time migration rates}

\item{fit_mig_mat}{(optional) Matrix of forwards-in-time migration rates}

\item{cr_mode}{Mode of prior on coalescent rates}

\item{cr_var}{Variance of prior on coalescent rates}

\item{mm_mode}{Mode of prior on backwards-in-time migration rates}

\item{mm_var}{Variance of prior on backwards-in-time migration rates}

\item{output_dir}{Directory to output log, tree and freq files to}

\item{run_name}{Prefix for log, tree and freq files. Outputs as 'run_name'.log etc}

\item{thin}{Thinning rate for MCMC samples}

\item{prop_rates}{Relative frequencies for each proposal type in the ratio Local_DTA:coal_rate update:bit_mig_mat update}
}
\description{
Samples N migration histories under the DTA model deme fixed at all leaves using belief propagation
}
