% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_sampling.R
\name{local_DTA_rejection}
\alias{local_DTA_rejection}
\title{Local DTA Update for EED data structure}
\usage{
local_DTA_rejection(
  EED,
  coal_node,
  fit_mig_mat,
  node_indices,
  max_attempts = 1e+05
)
}
\arguments{
\item{EED}{EED representation of a phylogeny including migration history}

\item{fit_mig_mat}{Forward-in-time (relative) migration matrix for the phylogeny}

\item{time_scale}{Time scale for the migration matrix}

\item{selected_node}{Label for node proposal will be centered on (optional)}
}
\value{
Proposal ED
}
\description{
Updates a migration history using the DTA model for adjacent branches to a coalescent node via rejection sampling
}
