% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterXML.R
\name{master.xml}
\alias{master.xml}
\title{Generates xml for structured coalescent simulation using MASTER}
\usage{
master.xml(
  effective.pop,
  migration.matrix,
  leaf.data,
  save.path = ".",
  run.xml = FALSE
)
}
\arguments{
\item{effective.pop}{effective population size from which the sample is taken}

\item{migration.matrix}{matrix of migration rates between demes}

\item{save.path}{file path to output .xml file to}

\item{run.xml}{if TRUE, run generated .xml through BEAST}

\item{data}{nx3 matrix with first column giving the tip labels, second column the time at which the sample was taken and third column the initial deme of each sample point}
}
\value{

}
\description{
Generates a .xml file to be input to BEAST 2 for structured coalescent simulations using MASTER
}
