% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAST2_xml.R
\name{master_xml}
\alias{master_xml}
\title{Generates xml for structured coalescent simulation using MASTER}
\usage{
master_xml(coal_rate, bit_mig_mat, leaf_data, n_deme, xml_path, con = stdout())
}
\arguments{
\item{coal_rate}{vector of coalescent rates}

\item{bit_mig_mat}{matrix of backward-in-time migration rates}

\item{leaf_data}{nx3 matrix or data.frame with first column giving tip labels, second column sample times and third column the initial deme of each sample point}

\item{n_deme}{number of distinct demes}

\item{xml_path}{path to output nexus file}

\item{con}{A connection object or a character string for location to output xml file}
}
\value{
xml file
}
\description{
Generates a .xml file to be input to BEAST 2 for structured coalescent simulations using MASTER
}
