% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDDistance.R
\name{medoid.ED}
\alias{medoid.ED}
\title{Medoid ED}
\usage{
medoid.ED(ED_sample, plot_medoid = FALSE)
}
\arguments{
\item{ED_sample}{Sample of ED objects containing migration histories on the same topology}

\item{plot_medoid}{Logical whether to plot the medoid migration history}
}
\value{
Medoid ED
}
\description{
Computes the medoid migration history from a sample (minimal mean distance
from all other sampled migration histories)
}
