% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{mig.death}
\alias{mig.death}
\title{Migration Death MCMC Move}
\usage{
mig.death(edge, edge.deme, node.ages, n.deme)
}
\arguments{
\item{edge}{Edge matrix for a structured coalescent tree}

\item{edge.deme}{Label for the deme in which each edge lies}

\item{node.ages}{Time of each node in the coalescent tree}

\item{n.deme}{Number of possible demes in the process}
}
\value{
List containing the updated edge matrix,
node ages and edge demes
}
\description{
Performs a Migration death move (Ewing et al. 2004). Selects an ancestral node
uniformly at random and removes the parent of the selected node, allocating
demes as necessary for modified edges.
}
