% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{mig.pair.birth}
\alias{mig.pair.birth}
\title{Migration Pair Birth MCMC Move}
\usage{
mig.pair.birth(edge, edge.deme, node.ages, n.deme)
}
\arguments{
\item{edge}{Edge matrix for a structured coalescent tree}

\item{edge.deme}{Label for the deme in which each edge lies}

\item{node.ages}{Time of each node in the coalescent tree}

\item{n.deme}{Number of possible demes in the process}
}
\value{
List with 3 elements, first element the edge matrix, second element
the node ages and third element giving the edge demes
}
\description{
Performs a Migration pair birth move (Ewing et al. 2004). Adds two migration
nodes on an edge selected uniformly at random from a structured coalescent
process, allocating a deme for the added edge such that a migration event
does not target its origin deme.
}
