% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigRateUpdate.R
\name{mig.rate.update}
\alias{mig.rate.update}
\title{Migration Rates Update}
\usage{
mig.rate.update(ED, migration.matrix, n.deme = NA, alpha = 1, beta = 10)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{migration.matrix}{Current migration rates matrix for the structured coalescent process}

\item{n.deme}{Number of demes in the structured coalescent process (optional argument)}

\item{alpha}{Gamma shape parameter for the prior on migration rates}

\item{beta}{Gamma rate parameter for the prior on migration rates}
}
\value{
Updated migration rates matrix conditioned on the current migration history
}
\description{
Performs an update of the migration rates matrix via a Gibbs move with Gamma prior and Gamma proposal distribution
}
