% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{migration.history.mcmc}
\alias{migration.history.mcmc}
\title{Runs and MCMC on Migration Histories}
\usage{
migration.history.mcmc(
  N0 = 1e+05,
  N = 1e+06,
  ED,
  effective.pop,
  gen.length,
  migration.matrix,
  n.deme = NA,
  proposal.rates = c(rep(5, 4), 1, 1),
  eff.pop.prior.mean = 1,
  eff.pop.prior.var = 1,
  mig.prior.mean = 1/10,
  mig.prior.var = 1/20,
  likelihood = "Structured",
  output.plots = TRUE,
  output.folder = "./MCMC_Results"
)
}
\arguments{
\item{N0}{Number of iterations to use as a burn-in}

\item{leaf.deme}{vector of demes for each leaf in \code{phylo}}
}
\value{
An object of class \code{phylo} augmented with the deme of each node in the tree.
}
\description{
Runs an MCMC to infer migration histories under the structured coalescent model
}
