% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihoods.R, R/Local_likelihoods.R
\name{MTT_likelihood}
\alias{MTT_likelihood}
\alias{MTT_local_likelihood}
\title{MultiTypeTree Probability Density}
\usage{
MTT_likelihood(
  ED,
  bit_rates,
  ED_NI = NodeIndices(ED),
  eigen_vals = eigen(bit_rates)$values,
  eigen_vecs = eigen(bit_rates)$vectors,
  inverse_vecs = solve(eigen_vecs)
)

MTT_local_likelihood(
  ED,
  st_labels,
  bit_rates,
  ED_NI = NodeIndices(ED),
  eigen_vals = eigen(bit_rates)$values,
  eigen_vecs = eigen(bit_rates)$vectors,
  inverse_vecs = solve(eigen_vecs)
)
}
\arguments{
\item{ED}{Extended data representation of a structured genealogy}

\item{bit_rates}{Transition matrix corresponding to the backwards-in-time migration process}

\item{ED_NI}{Node indices obtained from NodeIndices()}

\item{eigen_vals}{Vector of eigenvalues of \code{bit_rates}}

\item{eigen_vecs}{Matrix of eigenvectors of \code{bit_rates}}

\item{inverse_vecs}{Inverse matrix of \code{eigen_vecs}}

\item{st_labels}{Reduced extended data structure consisting of the rows of ED corresponding to coalescent nodes in the selected subtree only}
}
\value{
List consisting of the log.likelihood and likelihood of the structured genealogy
}
\description{
Computes the relevant probability density to compute transition kernels for
the MultiTypeTree node retype move
}
