% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTypeTree.R
\name{MTT_node_retype}
\alias{MTT_node_retype}
\alias{MTT_node_retype_eigen}
\title{MultiTypeTree Node Retype}
\usage{
MTT_node_retype(ED, st_labels, bit_rates, NI = NodeIndicesC(ED))

MTT_node_retype_eigen(
  ED,
  st_labels,
  bit_rates,
  NI = NodeIndicesC(ED),
  eigen_vals = eigen(bit_rates)$values,
  eigen_vecs = eigen(bit_rates)$vectors,
  inverse_vecs = solve(eigen_vecs)
)
}
\arguments{
\item{ED}{Extended data representation of phylogenetic tree and initial migration history}

\item{st_labels}{Reduced extended data structure consisting of the rows of ED corresponding to coalescent nodes in the selected subtree only}

\item{bit_rates}{Transition matrix of a Markov process corresponding to a process with rates given by the backwards-in-time migration matrix}

\item{NI}{Vector of row indices corresponding to which row of ED corresponds to each node label}
}
\value{
List consisting of proposal, a structured phylogeny with updates made to the selected subtree, and prop_prob, the probability of the given update
}
\description{
Generates a new migration history on a subtree drawn using MTT_st_coal_node()
}
