% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTypeTree.R
\name{MTT_node_retype_MCMC}
\alias{MTT_node_retype_MCMC}
\title{MultiTypeTree MCMC}
\usage{
MTT_node_retype_MCMC(
  N = 1e+06,
  ED,
  coal_rate,
  bit_mig_mat,
  st_depth = 1,
  cr_mode = 1,
  cr_var = 1,
  mm_mode = 0.05,
  mm_var = 0.5,
  thin = 1000,
  prop_rates = c(100, 1, 1),
  output_dir = "~",
  run_name = "MTT_R"
)
}
\arguments{
\item{N}{Number of MCMC iterations}

\item{ED}{Extended data representation of phylogenetic tree and initial migration history}

\item{coal_rate}{Vector of initial coalescent rates for the MCMC chain}

\item{bit_mig_mat}{Matrix of initial backwards-in-time migration rates for the MCMC chain}

\item{st_depth}{Number of generations below a coalescent node to include in the subtree for the node retype move}

\item{cr_mode}{Mode for prior on coalescent rates}

\item{cr_var}{Variance for prior on coalescent rates}

\item{mm_mode}{Mode for prior on backwards-in-time migration rates}

\item{mm_var}{Variance for prior on backwards-in-time migration rates}

\item{thin}{Thinning increment for the MCMC chain. A sample is saved every \texttt{thin} iterations}

\item{prop_rates}{Relative rates for proposals of migration history updates to coalescent rate updates to backwards-in-time migration rates updates}

\item{output_dir}{Directory to output files to for (thinned) results from the MCMC chain}

\item{run_name}{Name for run in output_dir}
}
\value{
Output files output_dir/run_name.freq, output_dir/run_name.log and output_dir/run_name.trees giving the acceptance rates of each MCMC move, thinned samples of coalescent and migration rates, and thinned samples of trees respectively
}
\description{
Runs a MCMC chain drawing joint samples of migration histories, backwards-in-time migration rates and coalescent rates.
MCMC chain runs with the node retype operator from MultiTypeTree for migration history updates and Gamma Gibbs moves for migration and coalescent rates
}
