% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTypeTree.R
\name{MTT_proposal_like}
\alias{MTT_proposal_like}
\alias{MTT_proposal_like_eigen}
\alias{MTT_local_like}
\alias{MTT_local_like_eigen}
\title{MultiTypeTree Probability Density}
\usage{
MTT_proposal_like(ED, bit_mig_mat, NI = NodeIndicesC(ED))

MTT_proposal_like_eigen(
  ED,
  bit_rates,
  NI = NodeIndicesC(ED),
  eigen_vals = NULL,
  eigen_vecs = NULL,
  inverse_vecs = NULL
)

MTT_local_like(ED, st_labels, bit_mig_mat, NI = NodeIndicesC(ED))

MTT_local_like_eigen(
  ED,
  st_labels,
  bit_rates,
  NI = NodeIndicesC(ED),
  eigen_vals = NULL,
  eigen_vecs = NULL,
  inverse_vecs = NULL
)
}
\arguments{
\item{ED}{Extended data representation of phylogenetic tree and initial migration history}

\item{bit_mig_mat}{Matrix of initial backwards-in-time migration rates for the MCMC chain}

\item{NI}{Vector of row indices corresponding to which row of ED corresponds to each node label}

\item{st_labels}{Reduced extended data structure consisting of the rows of ED corresponding to coalescent nodes in the selected subtree only (local likelihood only)}
}
\value{
List consisting of the log.likelihood and likelihood of the structured genealogy
}
\description{
Computes the relevant probability density to compute transition kernels for
the MultiTypeTree node retype move
}
