% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTypeTree.R
\name{MTT_st_coal_node}
\alias{MTT_st_coal_node}
\title{MultiTypeTree Subtree}
\usage{
MTT_st_coal_node(ED, st_depth = 1, NI = NodeIndicesC(ED), selected_node = NA)
}
\arguments{
\item{ED}{Extended data representation of phylogenetic tree and initial migration history}

\item{st_depth}{Number of generations below the selected coalescent node to include in the subtree}

\item{NI}{Vector of row indices corresponding to which row of ED corresponds to each node label}

\item{selected_node}{(optional) Selected coalescent node to use as centre of subtree}
}
\value{
List consisting of ED, the structured phylogeny input, and st_labels, a reduced extended data structure consisting of the coalescent nodes in the subtree only
}
\description{
Extracts a subtree (and associated sub-migration history) of a structured
genealogy consisting of the parent branch of a coalescent event and a fixed
number of generations of descendant branches
}
