% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeAgePhylo.R
\name{node.age.phylo}
\alias{node.age.phylo}
\title{Phylo object from node ages}
\usage{
node.age.phylo(edge, edge.deme, node.ages, tip.label = NA, plot.phylo = FALSE)
}
\arguments{
\item{edge}{Edge matrix for a structured coalescent tree}

\item{edge.deme}{Label for the deme in which each edge lies}

\item{node.ages}{Time of each node in the coalescent tree}

\item{tip.label}{(optional) Label for each tip; if not entered, labeled 1:n.leaf}

\item{plot.phylo}{logical; if FALSE (default) plot is not produced}
}
\value{
An object of class \code{phylo} (from package \code{ape}) storing the given structured coalescent process
}
\description{
Produces a phylo object under a structured coalescent model using the age of
each node
}
