% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NodeCount}
\alias{NodeCount}
\title{NodeCountC}
\usage{
NodeCount(ED, n_deme, node_indices)
}
\arguments{
\item{ED}{NumericMatrix Extended data object representing structured phylogeny}

\item{n_deme}{int Number of demes modelled under ED}

\item{node_indices}{NumericVector Vector giving row numbers in ED for node labels}
}
\value{
List containing matrix m with element (i,j) giving the number of migrations i -> j backwards in time, and a vector c with element i giving the number of coalescences occurring in deme i
}
\description{
Counts the number of migrations between pairs of demes, and coalescences within each deme
}
