% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NodeIndices}
\alias{NodeIndices}
\title{Node Indices}
\usage{
NodeIndices(ED)
}
\arguments{
\item{ED}{NumericMatrix Extended data object representing structured coalescent genealogy}
}
\description{
Returns a vector giving the row indices of the labels in ED
}
