% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{parent.node}
\alias{parent.node}
\title{Parent node}
\usage{
parent.node(node, edge, node.ages)
}
\arguments{
\item{node}{Node to find parent node of}

\item{edge}{Edge matrix from a \code{phylo} object}

\item{node.ages}{Time of each node in the \code{phylo} object}
}
\value{
Node ID of the parent node. If the input node is the root, returns Inf
}
\description{
Given a node from a phylo object, identifies the node ID of the parent node
}
