% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marginal_likelihood_estimate.R
\name{partial_marginal_likelihood}
\alias{partial_marginal_likelihood}
\title{Partial marginal likelihood computation}
\usage{
partial_marginal_likelihood(ED, cr_mode, cr_var, mm_mode, mm_var, n_deme)
}
\arguments{
\item{ED}{Migration history in ED format (9 columns)}

\item{cr_mode}{Coalescent rates prior mode}

\item{cr_var}{Coalescent rates prior variance}

\item{mm_mode}{Backward-in-time migration rates prior mode}

\item{mm_var}{Backward-in-time migration rates prior variance}

\item{n_deme}{Number of demes}
}
\description{
Computes the partial marginal likelihood obtained by integrating coalescent and migration rates out of the structured coalescent posterior
}
