% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloLikelihood.R
\name{phylo.likelihood}
\alias{phylo.likelihood}
\title{Calculates Likelihood of a Coalescent Tree}
\usage{
phylo.likelihood(phylo, effective.pop, gen.length)
}
\arguments{
\item{phylo}{object of class \code{phylo}}

\item{effective.pop}{effective population size from which the sample is taken}

\item{gen.length}{mean generation length of the sampled individuals}
}
\value{
A vector (\code{log-likelihood}, \code{likelihood}) giving the log-likelihood and likelihood of \code{phylo}
}
\description{
Returns the likelihood and log-likelihood of a (possibly heterochronous)
coalescent tree
}
