% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED_conversion.R
\name{phylo.to.ed}
\alias{phylo.to.ed}
\title{Phylo Object to Extended Data}
\usage{
phylo.to.ed(phylo)
}
\arguments{
\item{phylo}{phylo object augmented with the deme of each node}
}
\value{
A matrix with each row containing the parent, children, deme and age of
each node in the tree.
}
\description{
Converts a phylo object for a structured coalescent tree into an extended
data format containing the parent and children of each node alongside each
node's deme
}
