% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihoods.R
\name{phylo_likelihood}
\alias{phylo_likelihood}
\title{Kingman Coalescent Likelihood}
\usage{
phylo_likelihood(phylo, coal_rate)
}
\arguments{
\item{phylo}{object of class \code{phylo}}

\item{coal_rate}{Coalescent rate of the population (reciprocal of the product of effective population size and generation length)}
}
\value{
A vector (\code{log-likelihood}, \code{likelihood}) giving the log-likelihood and likelihood of \code{phylo}
}
\description{
Returns the likelihood and log-likelihood of a (possibly heterochronous)
coalescent tree
}
