% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtree_selection.R
\name{radius_subtree}
\alias{radius_subtree}
\alias{coal_node_subtree}
\title{Subtree sampling}
\usage{
radius_subtree(
  ED,
  ED_NI = NodeIndices(ED),
  st_radius,
  st_child = NULL,
  st_centre_loc = NULL,
  edge_lengths = NULL
)

coal_node_subtree(
  ED,
  ED_NI = NodeIndices(ED),
  st_depth = 1,
  selected_node = NULL
)
}
\arguments{
\item{ED}{Extended data representation of a structured genealogy}

\item{ED_NI}{Node indices obtained from NodeIndices()}

\item{st_radius}{Maximum radius of the sampled subtree. If a leaf or the root are reached, radius may be smaller than st_radius}

\item{st_child}{(optional) Child node of the branch segment containing the subtree centre. If omitted, child node is selected uniformly at random}

\item{st_centre_loc}{(optional) Location of the subtree centre between st_child and st_child's parent. If omitted, sampled uniformly from (0,1)}

\item{edge_lengths}{(optional) Vector of lengths of branch segments between every node and its parent in the order of rows of ED}

\item{st_depth}{Maximum number of descendant generations of coalescent nodes to include in the subtree (default 1)}

\item{selected_node}{(optional) Coalescent node to center subtree on. If omitted, a coalescent node is selected uniformly at random}
}
\value{
List consisting of the structured genealogy with added self-migrations at the subtree root and leaves (if necessary) and ED object consisting only of subtree nodes (including any added self-migrations)
}
\description{
Methods to sample subtrees and associated sub-migration histories from
structured genealogies
}
\details{
\code{radius_subtree} gives a radius-based subtree where points are included in the subtree up to a maximum distance from a fixed centrepoint.

\code{coal_node_subtree} gives a coalescent node-based subtree consisting of the parent branch and child branches up to \code{st_depth} generations of coalescent nodes below a central coalescent node
}
