% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{rnorm.reflect}
\alias{rnorm.reflect}
\title{Normal Distribution with Reflecting Boundary}
\usage{
rnorm.reflect(n, mean = 0, sd = 1, lower = -Inf, upper = Inf)
}
\arguments{
\item{n}{Number of variates to generate}

\item{mean}{Mean of the normal distribution}

\item{sd}{Standard deviation of the normal distribution}

\item{lower}{Lower reflecting boundary (possibly infinite)}

\item{upper}{Upper reflecting boundary (possibly infinite)}
}
\description{
Generates normally distributed variates subject to reflecting boundary
conditions
}
