% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{sample.vector}
\alias{sample.vector}
\title{Vector sampling}
\usage{
sample.vector(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose,
    or a positive integer.  See \sQuote{Details.}}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements
    of the vector being sampled.}
}
\description{
Samples a vector in the same way as base::sample except if the vector has
length 1, the single value is always returned
}
