% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SC_like_C}
\alias{SC_like_C}
\title{ScaledLikelihoodC}
\usage{
SC_like_C(ED, coal_rate, bit_mig_mat, node_indices)
}
\arguments{
\item{ED}{NumericMatrix Extended data object representing structured phylogeny}

\item{coal_rate}{NumericVector Effective population of each deme}

\item{node_indices}{NumericVector Vector giving row numbers in ED for node labels}

\item{mig_mat}{NumericMatrix Backwards-in-time migration rates between pairs of demes}
}
\value{
List containing log-likelihood and likelihood of the structured coalescent genealogy
}
\description{
Computes the likelihood of a structured coalescent genealogy
}
