% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihoods.R, R/Local_likelihoods.R
\name{SC_likelihood}
\alias{SC_likelihood}
\alias{SC_local_like}
\title{Structured Coalescent Likelihood}
\usage{
SC_likelihood(ED, coal_rate, bit_mig_mat, ED_NI)

SC_local_like(
  ED,
  coal_rate,
  bit_mig_mat,
  st_labels,
  ED_NI = NodeIndices(ED),
  ED_DD = DemeDecomp(ED, nrow(bit_mig_mat), ED_NI),
  ED_NC = NodeCount(ED, nrow(bit_mig_mat), ED_NI)
)
}
\arguments{
\item{ED}{Extended data representation of a structured genealogy}

\item{coal_rate}{Vector of coalescent rates}

\item{bit_mig_mat}{Matrix of backwards-in-time migration rates}

\item{ED_NI}{Node indices obtained from NodeIndices()}

\item{st_labels}{Reduced extended data structure consisting of the rows of ED corresponding to nodes in the selected subtree only}

\item{ED_DD}{Deme decomposition obtained from DemeDecomp()}

\item{ED_NC}{Node count obtained from NodeCount()}
}
\value{
A vector (\code{log-likelihood}, \code{likelihood}) giving the log-likelihood and likelihood of \code{phylo}
}
\description{
Computes the likelihood and log-likelihood of a structured genealogy under the
structured coalescent model
}
