% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_likelihoods.R
\name{SC_local_like}
\alias{SC_local_like}
\title{Local Structured Coalescent Likelihood}
\usage{
SC_local_like(
  ED,
  coal_rate,
  bit_mig_mat,
  event_ids,
  node_indices = NULL,
  deme_decomp = NULL,
  node_count = NULL
)
}
\arguments{
\item{ED}{Extended data representation of a phylogeny including migration history}

\item{coal_rate}{Vector of coalescent rates}

\item{bit_mig_mat}{Backward-in-time migration matrix for the phylogeny}

\item{event_ids}{Pair of event IDs to compute likelihood between}

\item{node_indices}{Node indices}

\item{deme_decomp}{Deme Decomp}
}
\value{
Likelihood & log_likelihood between events
}
\description{
Computes the structured coalescent likelihood locally between two specified events
}
