% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rate_proposals_scaled.R
\name{scaled_coal_rate_gibbs_update}
\alias{scaled_coal_rate_gibbs_update}
\title{Coalescent Rate Gibbs Update}
\usage{
scaled_coal_rate_gibbs_update(
  ED,
  time_scale,
  n_deme,
  node_indices,
  prior_shape = 1,
  prior_rate = 1
)
}
\arguments{
\item{ED}{Extended data representation of a phylogeny including migration history}

\item{time_scale}{Time scale for the migration matrix}

\item{node_indices}{Vector identifying row of each node label}

\item{prior_shape}{Gamma shape parameter for prior}

\item{prior_rate}{Gamma rate parameter for prior}
}
\value{
Updated parameters
}
\description{
Performs Gibbs updates for all parameters in a structured coalescent phylogeny
}
