% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaledSim.R
\name{scaled_sim}
\alias{scaled_sim}
\title{Simulation of Heterochronous Structured Coalescent}
\usage{
scaled_sim(
  leaf_data,
  coal_rate,
  time_scale,
  n_deme,
  mig_mat,
  plot_phylo = FALSE
)
}
\arguments{
\item{leaf_data}{nx3 matrix with first column giving the tip labels, second column the time at which the sample was taken and third column the initial deme of each sample point}

\item{coal_rate}{effective population size from which the sample is taken}

\item{time_scale}{generation length of the sampled individuals}

\item{n_deme}{total number of demes}

\item{mig_mat}{matrix of migration rates between demes}

\item{plot_phylo}{logical; if FALSE (default) plot is not produced}
}
\value{
An object of class \code{phylo} (from package \code{ape}) augmented with the likelihood and log-likelihood of the simulated tree, and the deme of each node in the tree. Additional nodes are added to account for migration events between coalescences
}
\description{
Simulates a tree under the structured coalescent model
}
