% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rate_proposals_scaled.R
\name{scaled_time_scale_gibbs_update}
\alias{scaled_time_scale_gibbs_update}
\title{Time Scale Gibbs Update}
\usage{
scaled_time_scale_gibbs_update(
  coal_rate,
  mig_mat,
  ED,
  n_deme,
  node_indices,
  prior_shape = 0.001,
  prior_rate = 0.001
)
}
\arguments{
\item{coal_rate}{Vector of coalescent rates}

\item{mig_mat}{Matrix of migration rates}

\item{ED}{Extended data representation of a phylogeny including migration history}

\item{n_deme}{Number of demes}

\item{node_indices}{Vector identifying row of each node label}

\item{prior_shape}{Gamma shape parameter for prior}

\item{prior_rate}{Gamma rate parameter for prior}
}
\value{
Updated parameters
}
\description{
Performs Gibbs updates for all parameters in a structured coalescent phylogeny
}
