% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScaledDTALikelihoodC}
\alias{ScaledDTALikelihoodC}
\title{ScaledDTALikelihoodC}
\usage{
ScaledDTALikelihoodC(ED, coal_rate, time_scale, bit_mig_mat, node_indices)
}
\arguments{
\item{ED}{NumericMatrix Extended data object representing structured phylogeny}

\item{coal_rate}{NumericVector Coalescent rate in each deme (1 / effective population size)}

\item{bit_mig_mat}{NumericMatrix Backwards-in-time migration rates between pairs of demes}

\item{node_indices}{NumericVector Vector giving row numbers in ED for node labels}
}
\value{
List containing log-likelihood and likelihood of the structured coalescent genealogy
}
\description{
Computes the DTA likelihood of a structured genealogy
}
