% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ScaledLikelihoodC}
\alias{ScaledLikelihoodC}
\title{ScaledLikelihoodC}
\usage{
ScaledLikelihoodC(ED, coal_rate, time_scale, mig_mat, node_indices)
}
\arguments{
\item{ED}{NumericMatrix Extended data object representing structured phylogeny}

\item{mig_mat}{NumericMatrix Backwards-in-time migration rates between pairs of demes}

\item{node_indices}{NumericVector Vector giving row numbers in ED for node labels}

\item{eff_pop}{NumericVector Effective population of each deme}

\item{gen_length}{double Generation length of each individual in the global population}
}
\value{
List containing log-likelihood and likelihood of the structured coalescent genealogy
}
\description{
Computes the likelihood of a structured coalescent genealogy
}
