% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtree_centre.R
\name{st_centre_location}
\alias{st_centre_location}
\alias{st_centre_location2}
\alias{st_centre_location_unif}
\title{Subtree centre sampling}
\usage{
st_centre_location(ED, ED_NI, centre_age, root_row = which(is.na(ED[, 2])))

st_centre_location2(ED, ED_NI, centre_age, root_row = which(is.na(ED[, 2])))

st_centre_location_unif(ED, ED_NI)
}
\arguments{
\item{ED}{ED representation of a phylogeny including migration history}

\item{ED_NI}{NodeIndicesC(ED)}

\item{centre_age}{Age of the subtree centre (e.g. sampled uniformly between oldest tip and root)}

\item{root_row}{(optional) row of ED corresponding to the root of the tree}
}
\value{
List consisting of st_child, the label of the child event in ED and st_centre_loc giving the location in [0,1] along the branch segment above st_child where the subtree centre is located
}
\description{
Methods to identify the subtree centre for MCMC
}
\details{
st_centre_location and st_centre_location2 sample a lineage at time centre_age uniformly.
st_centre_unif selects a subtree centre location uniformly on the tree
}
