% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initial_tree.R
\name{strip.history}
\alias{strip.history}
\title{strip.history}
\usage{
strip.history(ED, node_indices = NA)
}
\arguments{
\item{ED}{Extended data representation of a phylogeny including migration history}
}
\value{
An object of class \code{ED} with only coalescent nodes remaining and all located in deme 1
}
\description{
Strips the migration history from a phylogenetic tree
}
