% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{structured.plot}
\alias{structured.plot}
\title{Plot Structured Coalescent Tree}
\usage{
structured.plot(phylo, n.demes = NA)
}
\arguments{
\item{phylo}{phylo object augmented with node demes}
}
\description{
Plots a structured coalescent tree input as a phylo object augmented with node demes
}
