% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuredSim.R
\name{Structured.sim}
\alias{Structured.sim}
\title{Simulation of Heterochronous Structured Coalescent}
\usage{
Structured.sim(
  data,
  effective.pop,
  gen.length,
  n.deme,
  migration.matrix,
  plot.phylo = FALSE
)
}
\arguments{
\item{data}{nx3 matrix with first column giving the tip labels, second column the time at which the sample was taken and third column the initial deme of each sample point}

\item{effective.pop}{effective population size from which the sample is taken}

\item{gen.length}{generation length of the sampled individuals}

\item{n.deme}{total number of demes}

\item{migration.matrix}{matrix of migration rates between demes}

\item{plot.phylo}{logical; if FALSE (default) plot is not produced}
}
\value{
An object of class \code{phylo} (from package \code{ape}) augmented with the likelihood and log-likelihood of the simulated tree, and the deme of each node in the tree. Additional nodes are added to account for migration events between coalescences
}
\description{
Simulates a tree under the structured coalescent model
}
