% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StructuredLikelihoodC}
\alias{StructuredLikelihoodC}
\title{StructuredLikelihoodC}
\usage{
StructuredLikelihoodC(ED, eff_pop, gen_len, mig_mat, node_indices)
}
\arguments{
\item{ED}{NumericMatrix Extended data object representing structured phylogeny}

\item{eff_pop}{NumericVector Effective population of each deme}

\item{mig_mat}{NumericMatrix Backwards-in-time migration rates between pairs of demes}

\item{node_indices}{NumericVector Vector giving row numbers in ED for node labels}

\item{gen_length}{double Generation length of each individual in the global population}
}
\value{
List containing log-likelihood and likelihood of the structured coalescent genealogy
}
\description{
Computes the likelihood of a structured coalescent genealogy
}
