% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTree.R
\name{summary_tree}
\alias{summary_tree}
\title{Summary tree}
\usage{
summary_tree(ED_sample, locations, n_deme = NA)
}
\arguments{
\item{ED_sample}{Sample of migration histories on same topology}

\item{locations}{Vector of locations to compute pie charts}

\item{n_deme}{(optional) Number of demes}
}
\value{
Returns deme frequencies at each location
}
\description{
Generates a summary tree with pie charts at selected locations
}
