% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntheticLikelihood.R
\name{synth.likelihood}
\alias{synth.likelihood}
\title{Calculates synthetic likelihood for a coalescent tree with migration history}
\usage{
synth.likelihood(ED, effective.pop, rate, migration.matrix, node.indices)
}
\arguments{
\item{ED}{Extended data representation of a structured coalescent process}

\item{effective.pop}{Effective population sizes}

\item{rate}{Poisson rate for synthetic likelihood}

\item{migration.matrix}{Migration matrix}

\item{node.indices}{}
}
\value{
A vector (\code{log-likelihood}, \code{likelihood}) giving the log-likelihood and likelihood of \code{phylo}
}
\description{
Computes a synthetic likelihood for a coalescent tree with migration history
with a Poisson marginal distribution on the number of migrations
}
