% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.beast.multiple.R
\name{write.beast.multiple}
\alias{write.beast.multiple}
\title{write.beast.multiple}
\usage{
write.beast.multiple(
  ...,
  file = stdout(),
  translate = TRUE,
  thin = 1,
  initial = 0
)
}
\arguments{
\item{...}{either (i) a single object of class "strphylo"; (ii) multiple objects of class "strphylo" separated by commas; or (iii) a list of "strphylo" objects}

\item{file}{a file name specified either as a string; if left as default, outputs to standard output}

\item{translate}{logical; if TRUE, tip labels are replaced with tokens}

\item{thin}{integer from which the sample has been thinned}

\item{initial}{integer first index of sample}
}
\value{
None
}
\description{
Export multiple strphylo objects to BEAST NEXUS file
}
