% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryTree.R
\name{consensus_tree}
\alias{consensus_tree}
\title{Consensus tree}
\usage{
consensus_tree(ED_list, n_splits = 5, consensus_prob = 0.6, plot = TRUE)
}
\arguments{
\item{n_splits}{Number of splits per branch used to discretise the tree}

\item{consensus_prob}{Consensus probability to be exceeded to allow colouring}

\item{plot}{Logical value of whether to plot consensus tree}

\item{ED_sample}{Sample of migration histories on same topology}
}
\value{
Returns consensus tree in ED format
}
\description{
Generates a consensus tree by discretising branches and colouring each section if the consensus probability is sufficiently large
}
