% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DemeDecompC}
\alias{DemeDecompC}
\title{DemeDecomp}
\usage{
DemeDecompC(ED, n_deme, node_indices)
}
\arguments{
\item{x}{numeric vector blah blah}
}
\value{
mean(x) blah blah
}
\description{
Testing :)
}
