% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DemeDecompC}
\alias{DemeDecompC}
\title{DemeDecompC}
\usage{
DemeDecompC(ED, n_deme, node_indices)
}
\arguments{
\item{ED}{NumericMatrix Extended data object representing structured phylogeny}

\item{n_deme}{int Number of demes modelled under ED}

\item{node_indices}{NumericVector Vector giving row numbers in ED for node labels}
}
\value{
List containing matrix k giving deme decomposition with number of lineages in each deme during each time interval, vector event.times containing sorted event times from phylogeny and vector time.increments giving time increments between events during phylogeny
}
\description{
Decomposes an extended data object to give number of lineages in each deme during each time increment
}
