% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAST2_xml.R
\name{fixed_tree_xml}
\alias{fixed_tree_xml}
\title{Generates BEAST2 xml for MultiType Tree}
\usage{
fixed_tree_xml(
  strphylo,
  n_deme,
  coal_rate,
  bit_mig_mat,
  N = 1e+07,
  thin = 1000,
  con = stdout(),
  BEAST2_package = "MTT",
  run_name = "$(filebase)"
)
}
\arguments{
\item{strphylo}{structured phylo object giving initialisation condition for fixed tree run with MultiType Tree}

\item{coal_rate}{initial estimate for coalescent rates}

\item{N}{total number of MCMC iterations (including burn-in)}

\item{thin}{thinning increment for the MCMC}

\item{con}{A connection object or a character string}

\item{BEAST2_package}{Select package to prepare xml file for, either MTT (MultiTypeTree) or BASTA}

\item{run_name}{Name for logger files to be saved as}

\item{bit_mig_rate}{initial estimate for backward-in-time migration matrix}
}
\value{
output file or file content on screen
}
\description{
Generates a .xml file to run BEAST2 using either MultiTypeTree or BASTA with a fixed tree.
Initialises with all coalescent rates the same and all backward-in-time migration rates the same
}
