% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_sampling.R
\name{local_DTA}
\alias{local_DTA}
\title{Local DTA Update}
\usage{
local_DTA(ED, mig_mat, time_scale, selected_node = NA)
}
\arguments{
\item{ED}{Extended data representation of a phylogeny including migration history}

\item{mig_mat}{Forward-in-time migration matrix for the phylogeny}

\item{time_scale}{Time scale for the migration matrix}

\item{selected_node}{Label for node proposal will be centered on (optional)}
}
\value{
Proposal ED
}
\description{
Updates a migration history under the DTA model locally around a single coalescent node
}
