% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMoves.R
\name{mig.pair.death}
\alias{mig.pair.death}
\title{Migration Pair Death MCMC Move}
\usage{
mig.pair.death(edge, edge.deme, node.ages, n.deme)
}
\arguments{
\item{edge}{Edge matrix for a structured coalescent tree}

\item{edge.deme}{Label for the deme in which each edge lies}

\item{node.ages}{Time of each node in the coalescent tree}

\item{n.deme}{Number of possible demes in the process}
}
\value{
List containing the updated edge matrix,
node ages and edge demes
}
\description{
Performs a Migration pair death move (Ewing et al. 2004). Deletes two
migration nodes if they lie between two edges in the same deme.
}
