% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeCount.R
\name{node.count}
\alias{node.count}
\title{Counting of Migration and Coalescent Events}
\usage{
node.count(phylo)
}
\arguments{
\item{phylo}{An object of class \code{phylo} augmented with a deme for each node in the tree (with the deme of an edge being given by the end closest to the tips of the tree)}
}
\value{
Returns a matrix M with entries \deqn{m_{ij}} \deqn{p_{x}(1)=Pr(y=1|x)}
}
\description{
Counts the number of migration events between each pair of demes and the number
of coalescence events occurring in each deme
}
