% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marginal_likelihood_estimate.R
\name{structured_coalescent_marginal_likelihood}
\alias{structured_coalescent_marginal_likelihood}
\title{Marginal likelihood estimation}
\usage{
structured_coalescent_marginal_likelihood(
  N,
  N0,
  method,
  ED,
  coal_rate,
  bit_mig_mat,
  st_radius,
  adaptive = TRUE,
  cr_mode,
  cr_var,
  mm_mode,
  mm_var,
  adaptation_rate = 0.6,
  target_accept_rate = 0.234,
  n_stones = 100,
  output_file = NULL
)
}
\arguments{
\item{N}{MCMC iterations to use in marginal likelihood estimation}

\item{N0}{MCMC iterations of burn-in from the posterior distribution}

\item{ED}{Initial migration history in ED format (9 columns)}

\item{coal_rate}{Initial coalescent rates}

\item{bit_mig_mat}{Initial backwards-in-time migration rates}

\item{st_radius}{Initial radius for migration history proposals}

\item{adaptive}{Logical value indicating whether to adapt the radius during the burn-in phase (default TRUE)}

\item{cr_mode}{Coalescent rates prior mode}

\item{cr_var}{Coalescent rates prior variance}

\item{mm_mode}{Backward-in-time migration rates prior mode}

\item{mm_var}{Backward-in-time migration rates prior variance}

\item{adaptation_rate}{Adaptation rate (unused if adaptive = FALSE)}

\item{target_accept_rate}{Target acceptance rate for migration history updates}

\item{n_stones}{Number of stepping stones to use (unused in harmonic mean or thermodynamic integration estimates)}
}
\description{
Estimates the marginal likelihood under the structured coalescent using either the harmonic mean, thermodynamic integration or stepping stones sampling
}
